+++ eddai_EliteDangerousApiInterface/eddn/service/__init__.py
@@ -1,0 +1,1 @@
+from eddn.service.Client import EddnClient
