+++ eddai_EliteDangerousApiInterface/eddn/service/dataAnalytics/__init__.py
@@ -1,0 +1,2 @@
+from eddn.service.dataAnalytics.Commodity3Analytic import Commodity3Analytic
+from eddn.service.dataAnalytics.JournalAnalytic import JournalAnalytic
