+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/BaseSerializer.py
