+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/SAASignalsFound/SAASignalsFoundHotspotSerializers.py
