+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/SAASignalsFound/SAASignalsFoundSerializers.py
