+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/SAASignalsFound/SAASignalsFoundSignalAndSampleSerializers.py
