+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/SAASignalsFound/__init__.py
