+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/__init__.py
@@ -1,0 +1,2 @@
+from eddn.service.serializer.journals.Docked import DockedSerializer
+from eddn.service.serializer.journals.FSDJumpSerializer import FSDJumpSerializer
