+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/scan/AtmosphereComponentSerializer.py
