+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/scan/BaseScanSerializer.py
