+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/scan/MaterialsSerializer.py
