+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/scan/PlanetScanSerializer.py
