+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/journals/scan/StarScanSerializer.py
