+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/nestedSerializer/BaseSecondarySerializer.py
@@ -1,0 +1,4 @@
+from eddn.service.serializer.BaseSerializer import BaseSerializer
+
+class BaseNestedSerializer(BaseSerializer):
+    pass
