+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/nestedSerializer/minorFaction/BaseMinorFactionSerializer.py
@@ -1,0 +1,7 @@
+from eddn.service.serializer.nestedSerializer.BaseSecondarySerializer import BaseNestedSerializer
+from rest_framework import serializers
+
+class BaseMinorFactionSerializer(BaseNestedSerializer):
+    Name = serializers.CharField(
+        min_length=1,
+    )
