+++ eddai_EliteDangerousApiInterface/eddn/service/serializer/nestedSerializer/minorFaction/__init__.py
@@ -1,0 +1,2 @@
+from eddn.service.serializer.nestedSerializer.minorFaction.BaseMinorFactionSerializer import BaseMinorFactionSerializer
+from eddn.service.serializer.nestedSerializer.minorFaction.MinorFactionInSystemSerializer import MinorFactionInSystemSerializer
