--- eddai_EliteDangerousApiInterface/eddn/service/seriallizers/BaseSecondarySerializer.py
@@ -1,36 +1,0 @@
-from eddn.service.seriallizers.BaseSerializer import BaseSerializer
-
-class BaseSecondariSerializer(BaseSerializer):
-    def save(self, **kwargs):
-        assert hasattr(self, '_errors'), (
-            'You must call `.is_valid()` before calling `.save()`.'
-        )
-
-        assert not self.errors, (
-            'You cannot call `.save()` on a serializer with invalid data.'
-        )
-
-        # Guard against incorrect use of `serializer.save(commit=False)`
-        assert 'commit' not in kwargs, (
-            "'commit' is not a valid keyword argument to the 'save()' method. "
-            "If you need to access data before committing to the database then "
-            "inspect 'serializer.validated_data' instead. "
-            "You can also pass additional keyword arguments to 'save()' if you "
-            "need to set extra attributes on the saved model instance. "
-            "For example: 'serializer.save(owner=request.user)'.'"
-        )
-
-        assert not hasattr(self, '_data'), (
-            "You cannot call `.save()` after accessing `serializer.data`."
-            "If you need to access data before committing to the database then "
-            "inspect 'serializer.validated_data' instead. "
-        )
-
-        validated_data = {**self.validated_data, **kwargs}
-
-        self.instance = self.update_or_create(validated_data)
-        assert self.instance is not None, (
-            '`update_or_create()` did not return an object instance.'
-        )
-
-        return self.instance
