--- eddai_EliteDangerousApiInterface/eddn/service/seriallizers/customFields/CustomChoiceField.py
@@ -1,81 +1,0 @@
-from django.utils.translation import gettext_lazy as _
-from rest_framework import serializers
-from django_filters.filters import EMPTY_VALUES
-from ed_station.models.Station import Station
-import re
-
-
-class CustomChoiceField(serializers.ChoiceField):
-    def to_internal_value(self, data):
-        if not data in EMPTY_VALUES:
-            reg = re.findall(r'_([a-zA-Z]{1,});$', data)
-            data = str(reg[0]) if reg else ''
-            data = data if data != 'None' else ''
-        return super().to_internal_value(data)
-
-class HappinessChoiceField(serializers.ChoiceField):
-    default_error_messages = {
-        'invalid_choice': _('"{input}" is not a valid choice.'),
-        'incorrect_format': _('"{input}" Incorrect format'),
-        'state_phase': _('"{input}" is not a valid phase state'),
-    }
-
-    def to_internal_value(self, data):
-        if data.__class__ != dict:
-            if data == '' and self.allow_blank:
-                return ''
-            regex = r'Faction_([a-zA-Z]{1,})Band([0-9]);$'
-            if re.match(regex, data):
-                self.fail('incorrect_format', input=data)
-        
-            reg = re.findall(regex, data)
-            happi = str(reg[0][0]) if reg else None
-            statePhase = int(reg[0][1]) if reg else None
-        else:
-            happi, statePhase = data.get('State'), data.get('StatePhase')
-
-        if not 0 < statePhase < 4:
-            self.fail('state_phase', input=statePhase)
-        
-        try:
-            return {
-                'State': self.choice_strings_to_values[str(happi)],
-                'StatePhase': statePhase,
-            }
-        except KeyError:
-            self.fail('invalid_choice', input=happi)
-
-    def to_representation(self, value):
-        if value in EMPTY_VALUES:
-            return value
-        return f"$Faction_{value.get('State')}Band{value.get('StatePhase')};"
-
-class RingClassmChoiceField(serializers.ChoiceField):
-    def to_internal_value(self, data):
-        if (not data in EMPTY_VALUES) and (not data in  self.choice_strings_to_values.keys()):
-            reg = re.findall(r'eRingClass_([a-zA-Z]{1,})$', data)
-            data = str(reg[0]) if reg else ''
-        return super().to_internal_value(data)
-
-class ReserveLevelChoiceField(serializers.ChoiceField):
-    def to_internal_value(self, data):
-        if str(data).endswith('Resources'):
-            data = data[:-9]
-        return super().to_internal_value(data)
-    
-class LandingPadsChoiceField(serializers.ChoiceField):
-    def to_internal_value(self, data: dict):
-        choices = Station.LandingPadChoices.__empty__
-
-        if data.__class__ != dict:
-            if data in EMPTY_VALUES:
-                return choices
-            
-        if data.get(Station.LandingPadChoices.Large.name, 0) > 0:
-            choices = Station.LandingPadChoices.Large
-        elif data.get(Station.LandingPadChoices.Medium.name, 0) > 0:
-            choices = Station.LandingPadChoices.Medium
-        elif data.get(Station.LandingPadChoices.Small.name, 0) > 0:
-            choices = Station.LandingPadChoices.Small
-
-        return choices
