+++ eddai_EliteDangerousApiInterface/eddn/tasks/__init__.py
@@ -1,5 +1,5 @@
 from eddai_EliteDangerousApiInterface.celery import app as celery_app
-from eddn.service.client import EddnClient
+from eddn.service import EddnClient
 from eddn.tasks.tasks import *
 
 celery_app.register_task(EddnClient)
